refactor: split structured output schema validation
